MB-67589: Fix MSVC warning C4146...
unary minus operator applied to unsigned type, result still unsigned

Reviewed-on: https://review.couchbase.org/c/kv_engine/+/235973
Tested-by: Build Bot <build@couchbase.com>

+++ engines/ep/src/collections/flush_accounting.cc
@@ -149,7 +149,7 @@ bool FlushAccounting::StatisticsUpdate::remove(
 
     if (compactionCallbacks == CompactionCallbacks::AnyRevision &&
         isDelete == IsDeleted::Yes) {
-        updateDiskSize(-oldSize);
+        updateDiskSize(-static_cast<ssize_t>(oldSize));
         return false;
     }
